requirements.txt
aiofiles==24.1.0
aiosmtplib==3.0.2
aiosqlite==0.16.1
aniso8601==8.0.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
atlastk==0.13.4
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==7.1.2
colorama==0.4.6
Cython==3.1.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-mail==1.4.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
h11==0.16.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
itsdangerous==1.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.1
orjson==3.10.18
passlib==1.7.4
pillow==11.2.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pypika-tortoise==0.5.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2020.1
PyYAML==6.0.2
requests==2.32.3
six==1.15.0
sniffio==1.3.1
SQLAlchemy==1.3.18
starlette==0.46.2
tortoise==0.1.1
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==1.0.1
